possibly insoluble problem
This was an attempt to factor out 'cache' so it could be used, for
instance, to cache fragments that run over a whole directory; like rss,
navigation pages, etc.

It rapidly becomes clear that we can't consider it a simple pipeline any
more - we have to specify multiple files on the command line. However,
we assume that every script only emits a single file.

BUT what if that process fails?

The conceit of this tool is that you can pass the "input" along to the
next script. But... we don't have a mechanism to pass multiple sources
along.

This suggests that at minimum, permaweb must be a single stdin>stdout
stream

We might have cache utilities that can do other things, but we must
define conventions for html pipeline differently than rss-making script.