refactor not working...
so the idea here was to make it possible to cache the results
of iterating over an entire directory - for rss, etc.

so I refactored 'cache' out to its own thing, and tried to make
permaweb call that. This changed the calling conventions for ALL
the scripts (used one test to explore that) to not accept input on
stdin, but from the file arguments. This would have unified it
with the RSS case, where you pass, for instance, the 10 most recently
updated files.

BUT the conceit of permaweb is that it's a pipeline where you can
skip over things

what does it mean if the RSS script fails? You have no output. You can't pass the file
args along to the next thing, which would have (for instance) validated
the content.

Perhaps rethink this. We could have fragment caching things as
utilities, but permaweb should still think in single-file-in, single-file-out?